Adjust ssize_t definition for 32-bit to long.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@4311 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-headers/include/rpcndr.h
@@ -78,7 +78,7 @@ extern "C" {
 #ifdef _WIN64
   __MINGW_EXTENSION typedef __int64 ssize_t;
 #else
-  typedef int ssize_t;
+  typedef long ssize_t;
 #endif
 #endif
